LayoutTests/imported/w3c/web-platform-tests/html/anonymous-iframe/local-storage-initial-empty-document.tentative.https.window-expected.txt


FAIL Local storage is correctly partitioned with regards to credentialless iframe in initial empty documents. assert_equals: Credentialless iframe can't access credentialled context expected "" but got "value_E"
